Prototype + module tests

+++ src/main/kotlin/ru/romanow/product/holder/repository/ProductRepository.kt
@@ -3,4 +3,4 @@ package ru.romanow.product.holder.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import ru.romanow.product.holder.entity.Product
 
-interface ProductRepository : JpaRepository<Product, Int>
+interface ProductRepository : JpaRepository<Product, Long>
